[scala]: Complete analysis package refactoring
