--- core/src/main/java/com/github/fburato/highwheelmodules/core/analysis/Analyser.java
@@ -1,5 +1,0 @@
-package com.github.fburato.highwheelmodules.core.analysis;
-
-public interface Analyser {
-    AnalyserModel.AnalysisResult analyse(AnalysisState state);
-}
