+++ core/src/main/scala/com/github/fburato/highwheelmodules/core/analysis/Analyser.scala
@@ -1,0 +1,6 @@
+package com.github.fburato.highwheelmodules.core.analysis
+
+
+trait Analyser {
+  def analyse(state: AnalysisState): AnalysisResult
+}
