+++ core/src/main/scala/com/github/fburato/highwheelmodules/core/analysis/AnalyserException.scala
@@ -1,0 +1,9 @@
+package com.github.fburato.highwheelmodules.core.analysis
+
+case class AnalyserException(msg: String, cause: Exception) extends RuntimeException(msg, cause)
+
+object AnalyserException {
+  def apply(msg: String): AnalyserException = AnalyserException(msg, null)
+
+  def apply(cause: Exception): AnalyserException = AnalyserException(null, cause)
+}
